Fix style for add expenses
